requirements.txt
amqp==2.2.2
appnope==0.1.0
attrs==17.4.0
billiard==3.5.0.3
boto3==1.5.30
botocore==1.8.44
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.2.1
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.2.4
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-filter==2.0.0
django-model-utils==3.1.1
django-modeltranslation==0.13b1
django-money==0.14.3
django-nine==0.1.13
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-s3-collectstatic==1.0
django-sortedm2m==1.5.0
django-storages==1.6.5
djangorestframework==3.9.1
djangorestframework-camel-case==1.0.3
docutils==0.14
factory-boy==2.10.0
Faker==0.8.10
fakeredis==0.9.0
flake8==3.5.0
flake8-isort==2.5
gunicorn==19.7.1
idna==2.6
invoke==0.22.1
ipaddress==1.0.22
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10.1
jmespath==0.9.3
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
py-moneyed==0.7.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.3
pytest==3.4.0
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.6.1
python-env==1.0.0
python3-openid==3.1.0
pytz==2017.3
raven==6.5.0
redis==2.10.6
requests==2.22.0
requests-oauthlib==0.8.0
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tap.py==2.4
testfixtures==6.3.0
text-unidecode==1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
vine==1.1.4
wcwidth==0.1.7
whitenoise==3.3.1
